Deploying Classification Model
Deploying a Classification Model by first containerizing it in Docker, creating an API link in FastAPI, and using Minikube to orchestrate the deployment enables the model to run in a scalable, local Kubernetes environment. This setup allows for seamless management of containerized applications, making it easier to test and scale the model's deployment within a Kubernetes cluster before moving to a production environment.